Set up gradle
