+++ Data/Nah.Nah.txt
@@ -1,4 +1,1 @@
- 1. D | 1 | finish homework
- 2. D | 0 | prom | Dec 20 2024, 7:00 PM | 12 20 2024, 9:00 PM
- 3. D | 0 | dinner with friends | Sep 15 2025, 7:00 PM | 09 15 2025, 9:00 PM
- 4. D | 0 | start reading new book
+ 1. D | 0 | asjndfjsad
