--- src/main/java/Nah/Exceptions/InvalidFileValueException.java
@@ -1,7 +1,0 @@
-package Nah.Exceptions;
-
-public class InvalidFileValueException extends NahException{
-    public InvalidFileValueException() {
-        super(" File contain invalid value\n");
-    }
-}
