--- src/main/java/Nah/Exceptions/LackDescriptionException.java
@@ -1,7 +1,0 @@
-package Nah.Exceptions;
-
-public class LackDescriptionException extends NahException{
-    public LackDescriptionException(String mess) {
-        super(mess);
-    }
-}
