--- src/main/java/Nah/Exceptions/NahException.java
@@ -1,7 +1,0 @@
-package Nah.Exceptions;
-
-public class NahException extends Exception{
-    public NahException(String mess) {
-        super(mess);
-    }
-}
