+++ text-ui-test/EXPECTED.TXT
@@ -145,7 +145,7 @@ ________________________________________________________________________________
 
 ________________________________________________________________________________
 
-NAH!!! Nah.Data.Task 10 doesn't exist. Please give an number between 1 and 4
+NAH!!! nah.data.Task 10 doesn't exist. Please give an number between 1 and 4
 
 ________________________________________________________________________________
